+++ packages/json_rpc_codegen/lib/json_rpc_codegen.dart
@@ -13,6 +13,7 @@ export 'package:stream_channel/stream_channel.dart' show StreamChannel;
 
 export 'src/annotations/defaults.dart';
 export 'src/annotations/json_rpc.dart';
+export 'src/annotations/names.dart';
 export 'src/base/client_base.dart';
 export 'src/base/peer_base.dart';
 export 'src/base/server_base.dart';