+++ packages/json_rpc_codegen_builder/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Support the `@MethodName` and `@ParamName` annotations to override the
+JSON-RPC names of generated methods and named parameters
+
 ## 2.0.0 - 2026-07-30
 ### Changed
 - Rework the library with a new interface